+++ src/include/cpp-client/api/abstract.h
@@ -23,14 +23,11 @@ class Abstract {
 protected:
   Host host_;
   std::unique_ptr<IHTTP> http_;
-  int version_;
 
-  Abstract(IHTTP* http, int version) : http_(http), version_(version) {}
-  explicit Abstract(int version) : http_(makeHTTP()), version_(version) {}
+  Abstract() : http_(makeHTTP()) {}
+  explicit Abstract(IHTTP* http) : http_(http) {}
 
 public:
-  int version() const noexcept { return this->version_; };
-
   void setHost(const char* const newHost, int newPort) { this->host_.set(newHost, newPort); };
 };
 /**/